\name{rCppGem-package}
\alias{rCppGem-package}
\alias{rCppGem}
\docType{package}
\title{
\packageTitle{rCppGem}
}
\description{
\packageDescription{rCppGem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rCppGem}
\packageIndices{rCppGem}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rCppGem}

Maintainer: \packageMaintainer{rCppGem}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
